requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.6.1
imbalanced-learn==0.12.1
xgboost==2.1.4
matplotlib==3.10.0
pyyaml==6.0.1